feat: handle batch state vs sub proposal state
