+++ libs/i18n/src/locales/en/governance.json
@@ -163,9 +163,10 @@
   "created": "Created",
   "CreateProposalAndDownloadJSONToShare": "Create proposal and download JSON to share",
   "currently": "currently",
-  "currentlySetTo": "Currently expected to ",
+  "Currently expected to <0>pass</0>": "Currently expected to <0>pass</0>",
+  "Currently expected to <0>fail</0>": "Currently expected to <0>fail</0>",
   "Currently expected to pass: conditions met for {{count}} of {{total}} proposals": "Currently expected to pass: conditions met for {{count}} of {{total}} Proposals",
-  "Currently expected to fail: only {{count}} of {{total}} proposals are passing": "Currently expected to fail: only {{count}} of {{total}} proposals are passing",
+  "Currently expected to fail: {{count}} of {{total}} proposals are passing": "Currently expected to fail: {{count}} of {{total}} proposals are passing",
   "CurrentValue": "Current value",
   "dataIsIdentical": "Data is identical",
   "date": "Date",
@@ -290,7 +291,7 @@
   "Keep track of locked tokens in your wallet with the VEGA (VESTING) token.": "Keep track of locked tokens in your wallet with the $VEGA (VESTING) token.",
   "latestProposals": "Latest proposals",
   "learnMore": "Learn more",
-  "left to vote": "left to vote",
+  "{{time}} left to vote": "{{time}} left to vote",
   "Link transaction": "Link transaction",
   "liquidityComingSoon": "Liquidity rewards coming soon",
   "liquidityIntro": "You can read about our incentive program in this <linkToPost>blog post</linkToPost>.",
